fixes #40 fixes #45 prevent login replays
The strings passed to prepareForLogin and authenticate are now modified, so the state can be saved back to the user database. Once authenticated, it can't be used again. It doesn't actually work as a timeout, but a login attempt invalidates previous login attempts, and a valid authentication also invalidates re-use of that login.